Move MyApp to separate file
- vs code sneakily generated mac and ios files
