+++ lib/widgets/my_app.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key, required this.title}) : super(key: key);
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  int _itemCount = 20;
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: RefreshIndicator(
+        onRefresh: () => Future.delayed(
+            const Duration(seconds: 3), () => setState(() => _itemCount += 20)),
+        child: ListView.builder(
+            itemCount: _itemCount,
+            itemBuilder: (_, index) => SizedBox(
+                  height: 100,
+                  child: Card(
+                    child: Center(child: Text(index.toString())),
+                  ),
+                )),
+      ),
+    );
+  }
+}
